(rename) buttons styling and types file
