+++ app/src/components/common/button/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { ButtonCSS, ButtonComponentProps, ButtonProps } from '../../../common/styling_types'
+import { ButtonCSS, ButtonComponentProps, ButtonProps } from '../../../common/button_styling_types'
 
 interface ButtonComponentPropsLocal extends ButtonComponentProps {
   disabled?: boolean
