+++ app/src/components/common/three_box_comments/index.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import { ConnectedWeb3Context, useConnectedWeb3Context } from '../../../hooks/connectedWeb3'
 import { THREEBOX_ADMIN_ADDRESS, THREEBOX_SPACE_NAME } from '../../../common/constants'
 import { getLogger } from '../../../util/logger'
-import { ButtonCSS } from '../../../common/styling_types'
+import { ButtonCSS } from '../../../common/button_styling_types'
 
 const logger = getLogger('Component::ThreeBoxComments')
